requirements/run.txt
## direct dependencies
aiodns==1.1.1
aiohttp==2.2.5
alembic==0.9.9
async-timeout==2.0.1
asyncpg==0.16.0
attrs==18.1.0
git+https://github.com/BryanForbes/botus_receptus@b6ec78e8542d4b4e7f520d756ca23984007238ca#egg=botus_receptus[db]
cchardet==2.1.1
click==6.7
git+https://github.com/Rapptz/discord.py@607771c4f49a12d984d782883a580e7332359a10#egg=discord.py
mypy-extensions==0.3.0
psycopg2==2.7.5
SQLAlchemy==1.2.8
uvloop==0.10.1

## The following requirements were added by pip freeze:
chardet==3.0.4
idna==2.6
Mako==1.0.7
MarkupSafe==1.0
multidict==4.3.1
pycares==2.3.0
python-dateutil==2.7.2
python-editor==1.0.3
six==1.11.0
websockets==5.0.1
yarl==1.2.6
